Add support for continuous relaxation within optimize_acqf_mixed_alternating (#2635)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/2635

`optimize_acqf_mixed_alternating` utilizes local search to optimize discrete dimensions. This works well when there are a small number of values for the discrete dimensions but it does not scale well as the number of values increases. To address this, we have been transforming the high-cardinality dimensions in Ax and only passing in the low-cardinality dimensions as part of `discrete_dims`.
This diff adds support for using continuous relaxation for discrete dimensions that have more than `max_discrete_values` (configurable via `options`).

Also updates the optimizer to fall back to `optimize_acqf` if there are no discrete dimensions left. This is more user friendly than erroring out (particularly when used through Ax).

Reviewed By: Balandat

Differential Revision: D66239005

fbshipit-source-id: 0878115eb08ea75acb34ad8e891cf88393d4e36c