[Actions] Update publish website workflow
Dramatically simplify the process of deploying and versioning. We remove the shell script and perform the fewer steps from within github actions.

A big change here is the move to using docusaurus's built-in versioning system. Before, we were abusing docusaurus's system to take a snapshot of the entire website. For now we are only handling our md docs, but in future commits we will handle versioning tutorials and sphinx.