[Actions] Publish website using release branch
For a variety of reasons we cannot have the github actions bot commit changes directly to the `main`
branch, but we still want the process of creating new docusaurus versions to be automated in our
release process.

The approach here is to have a release branch called `docusaurus-versions` which will contain all
of the docusaurus data for versioning. The behavior is:
    1. We always build the website from the docusaurus-versions branch
    2. Before building the website we always sync docusaurus-versions with main
    3. For new releases our github workflow will create the new docusaurus version in the docusaurus-versions branch

The main branch would then only be responsible for tracking the latest documentation, and our
github workflows can continue to automate the release process from the docusaurus-versions branch without touching main.