Remove current_x validation from mixed optimizer (#2678)
Summary:
## Motivation

Removes the validation that `current_x` is within bounds from the `continuous_step` of the mixed optimizer. In some cases (e.g., with `Log` transform in Ax), the `post_processing_func` can bring the `current_x` slightly outside of the bounds, which leads to the optimization erroring out. The optimizer is capable of handling these outside of bounds starting points, so the validation doesn't provide much benefit.

Pull Request resolved: https://github.com/pytorch/botorch/pull/2678

Test Plan: Removed the test checking for the validation.

Reviewed By: esantorella

Differential Revision: D68276788

Pulled By: saitcakmak

fbshipit-source-id: caf82a867eb448b9fa19a083f0c5a29acdfd8f6f