`RobustRelevancePursuitSingleTaskGP` with specialized `fit_gpytorch_mll` (#2690)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/2690

This commit introduces an abstract `RobustRelevancePursuitModel` and `RobustRelevancePursuitSingleTaskGP`, a specific implementation of the abstract class. The main purpose of the new class is to provide an identical interface to a canonical `SingleTaskGP`, but automatically extend the likelihood with the `SparseOutlierGaussianLikelihood`, and toggle the Relevance Pursuit algorithm automatically through the marginal likelihood optimization via `fit_gpytorch_mll` by dispatching on the model type. This makes the model and algorithm easy to use.

Reviewed By: esantorella

Differential Revision: D68353582

fbshipit-source-id: 3b1308743a6e373d438260871c52b46de20d0d76