Remove hard-coded dtype from best_f buffers (#2725)
Summary:

Specifiying `dtype=float` (equivalent to `torch.float64`) causes issues if the user wants to use single precision. See https://github.com/pytorch/botorch/issues/2724

Removing the `dtype` argument from `torch.as_tensor` will lead to using the existing dtype if the input is a tensor, and using `torch.get_default_dtype()` if the input is a python float or a list of floats. If the input is a numpy array, the corresponding torch dtype is used (e.g. `torch.float64` for `np.float64`).

Reviewed By: esantorella

Differential Revision: D69121327