Update nonlinear_constraint_is_feasible to return a boolean tensor (#2731)
Summary:

`nonlinear_constraint_is_feasible` checks whether the constraint is feasible for the given (batch of) candidates. Previously, this returned a boolean, which was False if any element of the batch was infeasible. This diff updates it to return a boolean tensor that  shows whether each batch is feasible. Use cases are updated to comply with the new behavior.

I'll utilize this in a follow up diff to introduce a helper that evaluates the feasibility of all forms of constraints.

Reviewed By: dme65

Differential Revision: D69209007