Silence input normalization / standardization warnings with custom covariance modules
Summary:
The default covar modules are designed to work with normalized & standardized data. If the user has a custom covar module, they may need a different scaling of the data, which makes these warnings irrelevant.

This diff adds an option to restrict input data checks to NaNs only, and enables it whenever a custom covar module is given.

Differential Revision: D69613875