Silence input normalization / standardization warnings with custom covariance modules (#2745)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/2745

The default covar modules are designed to work with normalized & standardized data. If the user has a custom covar module, they may need a different scaling of the data, which makes these warnings irrelevant.

This diff adds an option to restrict input data checks to NaNs only, and enables it whenever a custom covar module is given.

Reviewed By: sdaulton

Differential Revision: D69613875

fbshipit-source-id: 8545752ae4f532a76c59983b963fbd6b873dde1c