Do not count STOPPED as a model fitting failure (#2767)
Summary:

Discovered this while debugging a CUDA test failure. I don't know why this was only raised when running on GPU but it was failing model fitting with status `STOPPED` (maybe it was slower and somehow timed out in `scipy_minimize`?).

Either way, we shouldn't count termination due to user specified criterion as a model fitting failure.

Reviewed By: Balandat

Differential Revision: D70994718