Do not count STOPPED as a model fitting failure (#2767)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/2767

Discovered this while debugging a CUDA test failure. I don't know why this was only raised when running on GPU but it was failing model fitting with status `STOPPED` (maybe it was slower and somehow timed out in `scipy_minimize`?).

Either way, we shouldn't count termination due to user specified criterion as a model fitting failure.

Reviewed By: Balandat

Differential Revision: D70994718

fbshipit-source-id: 310f45bf55ad013fcf7e16e965f11036b351bf63