Fix `custom_botorch_model_in_ax` tutorial
Recent changes on the Ax side broke this, fixing the breakage with minimal effort for now to get CI back to a green state.

With some of the Ax API changes in flight, we should revamp this tutorial more generally (and potentially not do this within botorch at all but in Ax instead).