Fix `custom_botorch_model_in_ax` tutorial (#2793)
Summary:
Recent changes on the Ax side broke this, fixing the breakage with minimal effort for now to get CI back to a green state.

With some of the Ax API changes in flight, we should revamp this tutorial more generally (and potentially not do this within botorch at all but in Ax instead).

Pull Request resolved: https://github.com/pytorch/botorch/pull/2793

Reviewed By: mpolson64

Differential Revision: D72173607

Pulled By: Balandat

fbshipit-source-id: 4f75d2baeba3ebed54184f896b9a8fa5b3d56237