Detach tensor in `gen_candidates_scipy` to avoid test failure due to new warning
Summary:
https://github.com/pytorch/pytorch/pull/143261 added a new warning to pytorch that gets raised when converting a tensor with gradients to a scalar without detaching it first.

This caused some test failues in botorch (which are likely a bit overzealous, but we can fix that separately). This makes a change in `gen_candidates_scipy` to fix this.

There are most likely other occurrences of this in botorch that we should fix, but here I'm quickly addressing the test failures first.

Reviewed By: saitcakmak

Differential Revision: D72286925

fbshipit-source-id: f1ddf2dc841d94c8b69589b0051cfc87f384be29