Test against python 3.13
Python 3.13 is out and used, so we should test against it.

This also require updating the minimum supported torch version in the install script. We should probably aim to bump the minium required pytorch version to 2.2.0 in new gpytorch/linear_operator/botorch releases so that we don't have to deal with the numpy<2.0 issues going forward.