Test against python 3.13 (#2814)
Summary:
Python 3.13 is out and used, so we should test against it.

This also require updating the minimum supported torch version in the install script.

Side note: We should probably aim to bump the minium required pytorch version to 2.2.0 in new gpytorch / linear_operator / botorch releases so that we don't have to deal with the numpy<2.0 issues going forward.

Pull Request resolved: https://github.com/pytorch/botorch/pull/2814

Reviewed By: saitcakmak

Differential Revision: D72731935

Pulled By: Balandat

fbshipit-source-id: 5fabed25ed5d66f10f3e1f491ef4e439facc1233