LogEI: select cache_root based on model support (#2820)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/2820

This diff has the input constructor for LogEI check if the model can support cache_root rather than defaulting it to True. Currently anytime you use LogEI with a multi-task model it spews a log warning that cache_root is being disabled. This removes that log message by changing the default to be smarter.

Reviewed By: SebastianAment

Differential Revision: D72667330

fbshipit-source-id: 304a1c673e31c79e5cefdc0449e2ada103350895