Put outcome transforms into `train` mode in model constructors (#2817)
Summary:
Fixes a bug where an outcome transform that is passed in `eval` mode is not put into `train` mode in the constructor of some `Model`s. As a result, the internal buffers of the transform (if applicable) are not updated and the outcome transform is not applied as intended.

This hasn't really been much of an issue in the past, since we typically either use a default transform (which we construct as needed and so is already in train mode) or we create a new `OutcomeTransform` object every time we construct a new `Model` instance.

Fixes https://github.com/pytorch/botorch/issues/2812

Pull Request resolved: https://github.com/pytorch/botorch/pull/2817

Reviewed By: saitcakmak

Differential Revision: D72766770

Pulled By: Balandat

fbshipit-source-id: 185dbd48f9ecf18db9413e9f0af5093883ca7407