Rename ax/models -> ax/generators (#2856)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/2856

X-link: https://github.com/facebook/Ax/pull/3807

This is a continuation of renaming of Models to Generators. In a previous diff, the classes have been renamed. This diff updates the directory / module names. Renaming of the attributes is the last remaining step. I will attempt that in a separate diff.

Reviewed By: Balandat

Differential Revision: D75161069