Batching L-BFGS-B (using as much as possible from scipy's implementation) (#2870)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/2870

This diff changes our acq optimization in the cases where we use L-BFGS-B to use separate L-BFGS-B states/optimizers for different batch elements.
This PR goes a little bit into the internals of scipy to make their FORTRAN/C code work as good as possible for us.
Basically, it calls the C-code in a for-loop for each batch item, instead of only calling it once for an aggregated function.

For this, I have edited their scipy/optimize/_lbfgsb_py.py file to handle batched functions as well. The fortran code is still running sequentially: just the evaluations are run in parallel.

We can adapt the batch sizes to be larger now and will likely get further speedups through that (they should be very sizable for PFNs, but also for GPs, esp. when on GPUs).

Why does this speedup things?

1. We converge faster, as each subproblem is simpler. I could see this be particularly pronounced for problems with small dimensionality.
2. When some sub-problem is not converged only that sub-problem is re-evaluated, saving total compute as iterations are quicker then.

How is this slower?

1. Based on my measurements it looks like scipy's LBFGS-B implementation is really fast for "larger" dimensionalities when it comes to step times, while we perform a for-loop there. It is much slower converging, though, as one can imagine with batching a problem like we do. This might be possible to speedup again via threading/multi-processing, but *beware* pytorch cuda tends to work *really* badly with multi-processing.

Differential Revision: D75977689