Allow batched fixed features in `gen_candidates_scipy` and `gen_candidates_torch` (#2893)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/2893

This is a PR that should enable batching for mixed optimization later. To enable it, we need to allow setting different fixed features for different initial conditions during optimizations.

We do this by allowing passing tensors of shape [b] or [b,q] to `gen_candidates_scipy` and for `gen_candidates_torch` for compatibility.

Differential Revision: D77043260