Batching Mixed Optimization (#2895)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/2895

So far, our optimization in mixed search spaces work on each restart separately and sequentially instead of batching them.
Here, we change this to batch the restarts, based on the new l-bfgs-b implementation that supports this.
This speeds up mixed search spaces a lot (depending on the problem around 3-4x speedups).

Reviewed By: esantorella

Differential Revision: D76517454

fbshipit-source-id: c2b3ec3bb2bbe8a18212e8247b67391ddba2941b