[GHA] Add workflow to test website build
__duplicate of https://github.com/facebook/Ax/pull/3769__

Before this there was nothing stopping us from merging website-breaking changes. This wouldn't get caught until our daily cron failed to build the website. Some of these errors can be caught by just running the website locally, but others require building the prod version of the website which involves merging in (but not deleting) the docusaurus-versions branch. It's much simpler if we can automate this as a workflow at PR creation.

I've configured this to run on PRs only when files directly relevant to the website build are modified (to reduce overhead for unrelated PRs). For the main branch this will be more thorough and run every time a PR is merged.