Add workflow to test website build (#2916)
Summary:
_duplicate of https://github.com/facebook/Ax/pull/3769_

Before this there was nothing stopping us from merging website-breaking changes. This wouldn't get caught until our daily cron failed to build the website. Some of these errors can be caught by just running the website locally, but others require building the prod version of the website which involves merging in (but not deleting) the docusaurus-versions branch. It's much simpler if we can automate this as a workflow at PR creation.

I've configured this to run on PRs only when files directly relevant to the website build are modified (to reduce overhead for unrelated PRs). For the main branch this will be more thorough and run every time a PR is merged.

Pull Request resolved: https://github.com/pytorch/botorch/pull/2916

Test Plan:
In https://github.com/facebook/Ax/pull/3769 I previously opened a separate PR to trigger this new workflow as a test:
- PR: https://github.com/facebook/Ax/pull/3770
  - Workflow: https://github.com/facebook/Ax/actions/runs/14938483763/job/41971282706?pr=3770

Reviewed By: saitcakmak

Differential Revision: D77952640

Pulled By: CristianLara

fbshipit-source-id: 26479c6a06ed36924fcaba7a9bb9279dcb296084