Added `reset` method to StoppingCriterion (#2927)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/2927

A custom StoppingCriterion currently does not work as intended when multiple restarts are used in fit_gpytorch_mll_(torch).

Due to its preservation of state, the previous fitting attempt's data will be saved, and generally, the criterion will immediately be satisfied on the new attempt, thus stopping fitting on the first iteration.

StoppingCriterion is now a Protocol with methods `__call__` and `reset`.

Reviewed By: Balandat

Differential Revision: D78343624

fbshipit-source-id: 6897a9c3f50f47fa7db4e9617e51438afd56d7b8