Add BaseTestProblem.is_minimization_problem property (#2949)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/2949

Most, but not all, BoTorch test functions are defined as minimization problems. Whether a function defined for minimization or maximization is currently only communicated via docstrings and it is not exposed in the API. This diff adds a property, `is_minimization_problem` that allows us to check for this in the code.

Looked through the test functions and updated the property for any problem that specifies itself as maximization problem in the docstring. Problems that don't have this information are assumed to be minimization problems (this will be verified for those that have optimizers defined in the next diff).

Reviewed By: esantorella

Differential Revision: D79562000

fbshipit-source-id: e86cb2afa5c3160e1920b634cc519c1ebc6d5e10