Fix AugmentedRosenbrock problem and expand testing for optimizers (#2950)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/2950

Fixes the bounds & the function value for AugmentedRosenbrock. Fixes https://github.com/pytorch/botorch/issues/2948

The function had [0, 5] bounds for the fidelity parameters, which the docstring specifies to be [0, 1]. Similarly, the docstring was specifying the optimizer as minimizer, but the function values were negated to make it a maximizer. Removed the negation and updated the bounds.

Reviewed By: esantorella

Differential Revision: D79562260

fbshipit-source-id: 8ff9268b083dd49d5592cd141f06b5d4d940a671