Add input constructor for LogCEI (#2973)
Summary:

Similar to LogPOF, the input constructor converts a constraints_tuple (A,b) to a dictionary of input constraints. Also infers best_f from a training set, similar to LogEI.

The input constructor does not throw an error if objective_index lies in the constraints, as this is already handled by the LogCEI constructor (this is included as a test case).

Differential Revision: D80183196