log values for HVKG using LogEHVI (#2974)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/2974

HKVG to output log values when _log=True using the LogEHVI value function.

Also adds a _log flag to InverseCostWeightedUtility to output utilities in log-space.

if _log=True, it assumes that
* current_value is in log-space
* cost_aware_utility outputs in log-space (raises an error if cost_aware_utility does not have a _log flag or if its _log=False).

Note that InverseCostWeightedUtility does a logarithmic transform for the inputted costs; assumes that inputted costs are in the original space. This is so that one does not need to make any direct modification to the cost fn, and because InverseCostWeightedUtility already does some pre-processing to the costs (e.g. clipping).

tldr: HKVG assumes all of its inputs are logged, but InverseCostWeightedUtility does not.

Reviewed By: SebastianAment

Differential Revision: D80263869

fbshipit-source-id: 95c29aac348a5541bf79de3a27208896936e424b