Add support for missing tasks in mtgp (#2960)
Summary:
X-link: https://github.com/facebookexternal/botorch_fb/pull/24

X-link: https://github.com/facebook/Ax/pull/4121

Pull Request resolved: https://github.com/pytorch/botorch/pull/2960

Currently, cross-validation in Ax fails when using a MTGP if there are multiple metrics and only some metrics have been observed for some tasks. This is a modeling problem, since the model is a ModelListGP and not all MTGPs in the list are required to have the same tasks. Hence when you pass in a test input, the model errors out if there are not observations from that task in the training data.

This avoids the error by mapping (optionally) mapping unexpected tasks to the "target task". This does not change the default behavior. For cross-validation in Ax, predictions are discarded if there are no observations for a given (task, metric) pair.

This will still error out in Ax if data for the target trial is missing.

Reviewed By: ItsMrLin

Differential Revision: D79812024