Condition_on_observations adds data to train_inputs when input_transforms are applied (#2990)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/2990

When applying input transforms (even with transform_on_train=True), conditioning on new data does not add it to train_inputs. This is unintuitive, as it is not clear why input transforms should disallow the user from adding observations to the training data this way.

Since condition_on_observations returns a model (and thus, the user has the flexibility to override the old model or not with the fantasy model), the data should always be added to training.

Fixes https://github.com/pytorch/botorch/issues/2533.

Simple notebook on why this doesn't make sense:
N7899660

Reviewed By: saitcakmak

Differential Revision: D80830741

fbshipit-source-id: 422abfa9d9d6212e52e15fa25e445a73d0b3dc42