Non-uniform model weights in EnsembleModel & EnsemblePosterior (#2993)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/2993

Pull Request resolved:  https://github.com/pytorch/botorch/pull/2993

Adds support for non-uniform model weights in EnsembleModel & EnsemblePosterior.

Why bother?
- Setting weights to be non-uniform for `MatheronPathModel` (e.g., [0, 0, 0, 1]) for different models in an ensemble (e.g. a `FullyBayesianSingleTaskGP`) allows the drawing of function samples from fully Bayesian Models.
- Also implemented modified sampling, since the rsample() method of the ensemble posterior would not be valid with batch shapes otherwise. Sampling occurs along the ensemble dimension only.
Indended use in benchmarking.

What this diff does not do:

Create a batchable rsample for EnsemblePosteriors. The current implementation does not work as intended, as it does not sample exclusively over the ensemble dimension and cannot handle different ensemble weights for different batches. Instead, it samples one model in the ensemble and applies it across batches. Attempting to implement this was very cumbersome, and broke a number of tests across the stack.

Reviewed By: Balandat

Differential Revision: D80728012

fbshipit-source-id: 6a2316862d98f0275a199b9b1442040b675337ef