Add worst known feasible value to constrained problems (#3016)
Summary:

Add the worst known feasible value to the constrained BoTorch test functions. This is useful when evaluating the performance of different optimization methods as this value can be assigned to all infeasible trials. This has the desirable property that any feasible trial has better performance than an infeasible trial.

Reviewed By: esantorella

Differential Revision: D82384518