Add MCAcquisition support to PFN (#3031)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/botorch/pull/3031

Makes all of the necessary changes to use MCAcquisitions, particularly qLogNoisyExpectedImprovement, with PFNModel and its BoundedRiemannPosterior, and to use PFN with multiple metrics in MBM. Including:

* Redo batch handling in PFNModel.posterior to do the right thing with the batch shapes sent in by Botorch MCAcquisition
* Change BoundedRiemannPosterior.rsample_from_base_samples to use N(0, I) base samples, and thus work with Botorch Samplers, and register the appropriate sampler.
* Add multi-batch handling to BoundedRiemannPosterior.icdf.
* MBM uses a ModelList if models are not GPs.

The PFNModel does not correctly handle q-batches yet. When MCAcquisition asks for a q-batch, it treats it the same as a t-batch, and in particular, the posteriors for each point in the q-batch are independent. This means that:
* Batch optimization does not work correctly. It will run and generate multiple points, but those points won't actually be properly conditioned on the earlier points in the batch.
* Handling of the incumbent best for EI in the noisy case is not correctly handled either, in that we are using an independent estimate for f_best rather than one jointly sampled. This is equivalent to using a plug-in estimate for f_best, which is the current behavior of DiscretizedExpectedImprovement.

Both of these will need to be addressed in the future, but I think this diff where things at least run end-to-end will be an easier starting point for that work.

Reviewed By: SamuelGabriel

Differential Revision: D79667144

fbshipit-source-id: 3eaad4f40e632e2f2c2848ff88429bfc83eb41c3