Add EnsembleMapSaasGP (#3035)
Summary:

Introduces a simple `EnsembleMapSaasGP` model that will replace `get_fitted_map_saas_ensemble` (which fits individual non-ensemble models and combines them into a fully Bayesian GP). The model internally is a batched `ExactGP`, which behaves just like a multi-output `SingleTaskGP`. The `posterior` method is overwritten to produce a `MixtureGaussianPosterior`, which retains the old behavior of the ensemble model.

The benefit of this model class is that it can be fit just like any other GP model, using `ExactMarginalLogLikelihood` and `fit_gpytorch_mll`. As such, it is fully compatible with Ax's MBM setup (as long as `allow_batched_models=False`).

Reviewed By: sdaulton

Differential Revision: D83701925