Improve initialization with continuous relaxation in optimize_acqf_mixed_alternating (#3041)
Summary:

`post_processing_func` is often used to apply Ax transforms to ensure that the generated points are aligned with the Ax search space. The transforms expects the values they are called with to be a member of the transformed search space. Since the initialization step already includes rounding operations, and the post processing function is already applied before returning from `optimize_acqf_mixed_alternating`, we can skip it at initialization.

Differential Revision: D83979491