Improve initialization with continuous relaxation in optimize_acqf_mixed_alternating (#3041)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/botorch/pull/3041

`post_processing_func` is often used to apply Ax transforms to ensure that the generated points are aligned with the Ax search space. The transforms expects the values they are called with to be a member of the transformed search space. Since the initialization step already includes rounding operations, and the post processing function is already applied before returning from `optimize_acqf_mixed_alternating`, we can skip it at initialization.

Not using `post_processing_func` in the initialization gets around the errors that may be raised by `post_processing_func` (due to the points not being in the specified search space), and will eliminate excessive fallback to random initialization that we see in Ax.

Reviewed By: dme65

Differential Revision: D83979491

fbshipit-source-id: 783d634ff24c159d1a46777e534c372da60974b4